common Header in all apps with app selector
