+++ web/packages/plugins/apps/analysis-platform/src/App.tsx
@@ -4,16 +4,20 @@ import { Layout } from 'antd'
 
 import Routes from './Routes'
 import { TApp, TLayout } from './Types'
-import Header from './components/Layout/Header'
+import { Header } from '@dmt/common'
 import { backgroundColorDefault } from './components/Design/Colors'
 import Content from './components/Layout/Content'
 import Menu from './components/Layout/Menu'
 
 const MainLayout = (props: TLayout) => {
-  const { heading, content, settings } = props
+  const { heading, content, settings, allApps } = props
   return (
     <>
-      <Header appName={settings.label} />
+      <Header
+        appName={settings.label}
+        urlPath={settings.urlPath}
+        allApps={allApps}
+      />
       <Layout style={{ background: backgroundColorDefault }}>
         <Menu appRootPath={settings.urlPath} />
         <Content settings={settings} heading={heading} content={content} />
@@ -23,7 +27,7 @@ const MainLayout = (props: TLayout) => {
 }
 
 export default (props: TApp): JSX.Element => {
-  const { settings } = props
+  const { settings, applications } = props
 
   return (
     <>
@@ -34,6 +38,7 @@ export default (props: TApp): JSX.Element => {
           key={route.path}
           render={() => (
             <MainLayout
+              allApps={applications}
               heading={route.heading}
               content={route.content}
               settings={settings}
