+++ web/packages/plugins/shared/common/src/components/index.ts
@@ -7,3 +7,4 @@ export * from './UiPluginSelector'
 export * from './UploadFileButton'
 export * from './NewEntityButton'
 export * from './Dialog'
+export * from './Header'
