+++ web/packages/plugins/shared/common/src/types.ts
@@ -27,3 +27,19 @@ export type TContainerImage = {
   version: string
   registryName: string
 }
+
+export type DmtSettings = {
+  name: string
+  label: string
+  tabIndex: number
+  hidden: boolean
+  visibleDataSources: any
+  type: string
+  description: string
+  packages: any
+  models: any
+  actions: any
+  file_loc: string
+  data_source_aliases: any
+  urlPath: string
+}
